backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
requests==2.32.5
feedparser==6.0.11
apscheduler==3.11.0
python-dateutil==2.9.0.post0

